# spring-openapi-generator/api-schema/src/main/java/com/sciencesakura/sample/api/model/UpdateItemStockRequest.java
package com.sciencesakura.sample.api.model;

import java.net.URI;
import java.util.Objects;
import com.fasterxml.jackson.annotation.JsonProperty;
import com.fasterxml.jackson.annotation.JsonCreator;
import org.openapitools.jackson.nullable.JsonNullable;
import java.io.Serializable;
import java.time.OffsetDateTime;
import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import io.swagger.v3.oas.annotations.media.Schema;


import java.util.*;
import jakarta.annotation.Generated;

/**
 * UpdateItemStockRequest
 */

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen")
public class UpdateItemStockRequest implements Serializable {

  private static final long serialVersionUID = 1L;

  private Double quantity;

  public UpdateItemStockRequest() {
    super();
  }

  /**
   * Constructor with only required parameters
   */
  public UpdateItemStockRequest(Double quantity) {
    this.quantity = quantity;
  }

  public UpdateItemStockRequest quantity(Double quantity) {
    this.quantity = quantity;
    return this;
  }

  /**
   * Get quantity
   * @return quantity
  */
  @NotNull 
  @Schema(name = "quantity", requiredMode = Schema.RequiredMode.REQUIRED)
  @JsonProperty("quantity")
  public Double getQuantity() {
    return quantity;
  }

  public void setQuantity(Double quantity) {
    this.quantity = quantity;
  }

  @Override
  public boolean equals(Object o) {
    if (this == o) {
      return true;
    }
    if (o == null || getClass() != o.getClass()) {
      return false;
    }
    UpdateItemStockRequest updateItemStockRequest = (UpdateItemStockRequest) o;
    return Objects.equals(this.quantity, updateItemStockRequest.quantity);
  }

  @Override
  public int hashCode() {
    return Objects.hash(quantity);
  }

  @Override
  public String toString() {
    StringBuilder sb = new StringBuilder();
    sb.append("class UpdateItemStockRequest {\n");
    sb.append("    quantity: ").append(toIndentedString(quantity)).append("\n");
    sb.append("}");
    return sb.toString();
  }

  /**
   * Convert the given object to string with each line indented by 4 spaces
   * (except the first line).
   */
  private String toIndentedString(Object o) {
    if (o == null) {
      return "null";
    }
    return o.toString().replace("\n", "\n    ");
  }
}